ai-server/requirements.txt
anyio==3.6.2
boto3==1.26.126
botocore==1.29.126
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
fastapi==0.89.1
filelock==3.12.0
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.7.0.68
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
priority==2.0.0
protobuf==4.22.3
psutil==5.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.5
pytz==2023.3
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
s3transfer==0.6.0
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.22.0
sympy==1.11.1
thop==0.1.1.post2209072238
toml==0.10.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
typing_extensions==4.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
wsproto==1.2.0
